fixed d3.js missing from precompile list

+++ lib/tasks/fake_thoughts.rake
@@ -1,32 +1,34 @@
-require 'faker'
+if Rails.env.development?
+  require 'faker'
 
-namespace :db do
-  desc "Fill database with sample data"
-  task populate: :environment do
-    Thought.delete_all
+  namespace :db do
+    desc "Fill database with sample data"
+    task populate: :environment do
+      Thought.delete_all
     
-    tags = [
-      "ruby-on-rails", "ruby", "sinatra", "heroku", "http",
-      "html", "css", "scss", "jquery", "coffeescript", "javascript",
-      "python", "php", "java", "mcgill", "osx", "installing",
-      "facebook", "amazon-s3", "configuration"
-    ].combination(3).to_a
+      tags = [
+        "ruby-on-rails", "ruby", "sinatra", "heroku", "http",
+        "html", "css", "scss", "jquery", "coffeescript", "javascript",
+        "python", "php", "java", "mcgill", "osx", "installing",
+        "facebook", "amazon-s3", "configuration"
+      ].combination(3).to_a
     
-    115.times do
+      115.times do
       
-      params = {
-        title: Faker::Lorem.words(Random.rand(3) + 5).join(" "),
-        blurb: Faker::Lorem.sentence,
-        content: Faker::Lorem.paragraphs(Random.rand(2) + 2).join("\n\n"),
-        author: 'aj.ostrow',
-      }
+        params = {
+          title: Faker::Lorem.words(Random.rand(3) + 5).join(" "),
+          blurb: Faker::Lorem.sentence,
+          content: Faker::Lorem.paragraphs(Random.rand(2) + 2).join("\n\n"),
+          author: 'aj.ostrow',
+        }
       
-      thought = Thought.new params
+        thought = Thought.new params
       
-      thought.tag_list = tags.sample.sample(Random.rand(3) + 1).join(", ")
+        thought.tag_list = tags.sample.sample(Random.rand(3) + 1).join(", ")
       
-      thought.save
+        thought.save
+      end
+    
     end
-    
   end
 end
